Move /agencies/participation to /participation/agencies
Noticed it was also missing some test. Ooops.
