+++ tests/functional/test_geo.py
@@ -14,9 +14,3 @@ class TestGeoEndpoint:
     def test_county_detail_endpoint(self, testapp, swagger):
         res = testapp.get('/geo/counties/39043')
         assert res.status_code == 200
-
-    def test_state_participation_endpoint(self, testapp, swagger):
-        res = testapp.get('/geo/states/NY/participation')
-        print(res.json['results'][0])
-        assert res.status_code == 200
-        validate_api_call(swagger, raw_request=res.request, raw_response=res)
